Separate most of the houses types code into distinct files
