+++ src/GUI/UI/Houses/Tavern.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "GUI/UI/UIHouses.h"
+#include "GUI/UI/UIHouseEnums.h"
+
+void TavernDialog();
