+++ src/GUI/UI/UIHouses.cpp
@@ -36,6 +36,13 @@
 #include "GUI/UI/UIShops.h"
 #include "GUI/UI/UIStatusBar.h"
 #include "GUI/UI/Houses/MagicGuild.h"
+#include "GUI/UI/Houses/Bank.h"
+#include "GUI/UI/Houses/Jail.h"
+#include "GUI/UI/Houses/Tavern.h"
+#include "GUI/UI/Houses/Temple.h"
+#include "GUI/UI/Houses/Training.h"
+#include "GUI/UI/Houses/Transport.h"
+#include "GUI/UI/Houses/MercenaryGuild.h"
 
 #include "Io/Mouse.h"
 #include "Io/KeyboardInputHandler.h"
@@ -61,92 +68,6 @@ Image *_591428_endcap = nullptr;
 
 void GenerateStandartShopItems();
 void GenerateSpecialShopItems();
-
-struct stru365_travel_info {
-    unsigned char uMapInfoID;
-    unsigned char pSchedule[7];
-    unsigned int uTravelTime; // In days.
-    int arrival_x;
-    int arrival_y;
-    int arrival_z;
-    int arrival_view_yaw;
-    unsigned int uQuestBit;  // quest bit required to set for this travel option
-                             // to be enabled; otherwise 0
-};
-stru365_travel_info transport_schedule[35] = {  // 004F09B0
-                                                // location name   days    x        y       z
-    { MAP_STEADWICK, { 1, 0, 1, 0, 1, 0, 0 }, 2, -18048, 4636, 833, 1536, 0 },  // for stable
-    { MAP_PIERPONT, { 0, 1, 0, 1, 0, 1, 0 }, 2, -2527, -6773, 1153, 896, 0 },
-    { MAP_TATALIA, { 1, 0, 1, 0, 1, 0, 0 }, 2, 4730, -10580, 320, 1024, 0 },
-    { MAP_HARMONDALE, { 0, 1, 0, 1, 0, 1, 0 }, 2, -5692, 11137, 1, 1024, 0 },
-    { MAP_DEYJA, { 1, 0, 0, 1, 0, 0, 0 }, 3, 7227, -16007, 2625, 640, 0 },
-    { MAP_BRAKADA_DESERT, { 0, 0, 1, 0, 0, 1, 0 }, 3, 8923, 17191, 1, 512, 0 },
-    { MAP_AVLEE, { 1, 0, 1, 0, 1, 0, 0 }, 3, 17059, 12331, 512, 1152, 0 },
-    { MAP_DEYJA, { 0, 1, 0, 0, 1, 0, 1 }, 2, 7227, -16007, 2625, 640, 0 },
-    { MAP_HARMONDALE, { 0, 1, 0, 1, 0, 1, 0 }, 2, -5692, 11137, 1, 1024, 0 },
-    { MAP_STEADWICK, { 1, 0, 1, 0, 1, 0, 0 }, 3, -18048, 4636, 833, 1536, 0 },
-    { MAP_PIERPONT, { 0, 1, 0, 1, 0, 1, 0 }, 2, -2527, -6773, 1153, 896, 0 },
-    { MAP_STEADWICK, { 1, 0, 1, 0, 1, 0, 1 }, 3, -18048, 4636, 833, 1536, 0 },
-    { MAP_HARMONDALE, { 0, 1, 0, 0, 0, 1, 0 }, 5, -5692, 11137, 1, 1024, 0 },
-    { MAP_STEADWICK, { 0, 1, 0, 1, 0, 1, 0 }, 2, -18048, 4636, 833, 1536, 0 },
-    { MAP_PIERPONT, { 0, 1, 0, 1, 0, 1, 0 }, 3, -2527, -6773, 1153, 896, 0 },
-    { MAP_DEYJA, { 0, 0, 1, 0, 0, 0, 1 }, 5, 7227, -16007, 2625, 640, 0 },
-    { MAP_TATALIA, { 0, 1, 0, 1, 0, 1, 0 }, 2, -2183, -6941, 97, 0, 0 },
-    { MAP_AVLEE, { 1, 0, 0, 0, 1, 0, 0 }, 4, 7913, 9476, 193, 0, 0 },
-    { MAP_EVENMORN_ISLE, { 0, 0, 0, 0, 0, 0, 1 }, 7, 15616, 6390, 193, 1536, QBIT_EVENMORN_MAP_FOUND },
-    { MAP_BRAKADA_DESERT, { 0, 0, 1, 0, 0, 0, 0 }, 6, 19171, -19722, 193, 1024, 0 },
-    { MAP_AVLEE, { 0, 1, 0, 1, 0, 1, 0 }, 3, 7913, 9476, 193, 0, 0 },
-    { MAP_BRAKADA_DESERT, { 1, 0, 1, 0, 0, 0, 0 }, 6, 19171, -19722, 193, 1024, 0 },
-    { MAP_TATALIA, { 1, 0, 1, 0, 1, 0, 0 }, 4, -2183, -6941, 97, 0, 0 },
-    { MAP_PIERPONT, { 0, 0, 0, 0, 0, 1, 0 }, 6, -709, -14087, 193, 1024, 0 },  // for boat
-    { MAP_STEADWICK, { 0, 0, 0, 0, 0, 0, 1 }, 6, -10471, 13497, 193, 1536, 0 },
-    { MAP_EVENMORN_ISLE, { 0, 1, 0, 1, 0, 0, 0 }, 1, 15616, 6390, 193, 1536, QBIT_EVENMORN_MAP_FOUND },
-    { MAP_BRAKADA_DESERT, { 0, 1, 0, 1, 0, 0, 0 }, 1, 19171, -19722, 193, 1024, 0 },
-    { MAP_STEADWICK, { 0, 1, 0, 1, 0, 1, 0 }, 2, -10471, 13497, 193, 1536, 0 },
-    { MAP_BRAKADA_DESERT, { 1, 0, 1, 0, 0, 0, 0 }, 4, 19171, -19722, 193, 1024, 0 },
-    { MAP_EVENMORN_ISLE, { 0, 0, 0, 0, 0, 0, 1 }, 5, 15616, 6390, 193, 1536, QBIT_EVENMORN_MAP_FOUND },
-    { MAP_AVLEE, { 0, 0, 0, 0, 1, 0, 0 }, 5, 7913, 9476, 193, 0, 0 },
-    { MAP_STEADWICK, { 0, 1, 0, 0, 0, 1, 0 }, 4, -10471, 13497, 193, 1536, 0 },
-    { MAP_PIERPONT, { 1, 0, 1, 0, 1, 0, 0 }, 3, -709, -14087, 193, 1024, 0 },
-    { MAP_TATALIA, { 0, 0, 0, 1, 0, 0, 0 }, 5, -2183, -6941, 97, 0, 0 },
-    { MAP_ARENA, { 0, 0, 0, 0, 0, 0, 1 }, 4, 3844, 2906, 193, 512, 0 }
-};
-
-unsigned char transport_routes[20][4] = {
-    { 0, 1, 1, 34 },         // HOUSE_STABLES_HARMONDALE
-    { 2, 3, 4, 5 },          // HOUSE_STABLES_STEADWICK
-    { 6, 7, 8, 8 },          // HOUSE_STABLES_TULAREAN_FOREST
-    { 9, 10, 10, 10 },       // HOUSE_STABLES_DEYJA
-    { 11, 11, 12, 12 },      // HOUSE_STABLES_BRACADA_DESERT
-    { 13, 13, 13, 13 },      // HOUSE_STABLES_TATALIA
-    { 14, 14, 15, 15 },      // HOUSE_STABLES_AVLEE
-    { 255, 255, 255, 255 },  // HOUSE_STABLES_61
-    { 255, 255, 255, 255 },  // HOUSE_STABLES_62
-    { 255, 255, 255, 255 },  // HOUSE_BOATS_EMERALD_ISLE
-    { 16, 17, 18, 19 },      // HOUSE_BOATS_ERATHIA
-    { 18, 20, 21, 21 },      // HOUSE_BOATS_TULAREAN_FOREST
-    { 22, 23, 24, 25 },      // HOUSE_BOATS_BRACADA_DESERT
-    { 22, 22, 23, 23 },      // HOUSE_BOATS_EVENMORN_ISLAND
-    { 255, 255, 255, 255 },  // HOUSE_BOATS_68
-    { 27, 28, 29, 30 },      // HOUSE_BOATS_TATALIA
-    { 31, 32, 33, 33 },      // HOUSE_BOATS_AVLEE
-    { 24, 24, 24, 24 },      // HOUSE_BOATS_71
-    { 255, 255, 255, 255 },  // HOUSE_BOATS_72
-    { 255, 255, 255, 255 }   // HOUSE_BOATS_73
-};
-
-IndexedArray<int, HOUSE_TRAINING_HALL_EMERALD_ISLE, HOUSE_TRAINING_HALL_STONE_CITY> trainingHallMaxLevels = {
-    {HOUSE_TRAINING_HALL_EMERALD_ISLE, 5},
-    {HOUSE_TRAINING_HALL_HARMONDALE, 15},
-    {HOUSE_TRAINING_HALL_ERATHIA, 25},
-    {HOUSE_TRAINING_HALL_TULAREAN_FOREST, 25},
-    {HOUSE_TRAINING_HALL_CELESTE, 200},
-    {HOUSE_TRAINING_HALL_PIT, 200},
-    {HOUSE_TRAINING_HALL_NIGHON, std::numeric_limits<int>::max()}, // no limit
-    {HOUSE_TRAINING_HALL_TATALIA, 50},
-    {HOUSE_TRAINING_HALL_AVLEE, 50},
-    {HOUSE_TRAINING_HALL_STONE_CITY, 100},
-};
 
 std::array<const HouseAnimDescr, 196> pAnimatedRooms = { {  // 0x4E5F70
     { "", 0x4, 0x1F4, BuildingType_Invalid, 0, 0 },
@@ -547,8 +468,6 @@ const std::array<ITEM_VARIATION, 28> shopArmr_variation_spc = {{
     { ITEM_TREASURE_LEVEL_5, { 35, 35, 38, 38 } },
     { ITEM_TREASURE_LEVEL_5, { 33, 33, 33, 33 } }
 }};
-
-std::vector<int> charactersTrainedLevels;
 
 std::array<std::string, 6> portraitClickLabel;
 
@@ -1230,179 +1149,6 @@ void OnSelectShopDialogueOption(DIALOGUE_TYPE option) {
     }
 }
 
-int GetTravelTimeTransportDays(int schedule_id) {
-    int travel_time = transport_schedule[schedule_id].uTravelTime;
-    if (window_SpeakInHouse->wData.val >= HOUSE_BOATS_EMERALD_ISLE) {
-        if (CheckHiredNPCSpeciality(Sailor)) travel_time -= 2;
-        if (CheckHiredNPCSpeciality(Navigator)) travel_time -= 3;
-        if (CheckHiredNPCSpeciality(Pirate)) travel_time -= 2;
-    } else {
-        if (CheckHiredNPCSpeciality(Horseman)) travel_time -= 2;
-    }
-    if (CheckHiredNPCSpeciality(Explorer)) --travel_time;
-    if (travel_time < 1) travel_time = 1;
-    return travel_time;
-}
-
-void TravelByTransport() {
-    GUIWindow travel_window = *window_SpeakInHouse;
-    travel_window.uFrameX = SIDE_TEXT_BOX_POS_X;
-    travel_window.uFrameWidth = SIDE_TEXT_BOX_WIDTH;
-    travel_window.uFrameZ = SIDE_TEXT_BOX_POS_Z;
-
-    assert(pParty->hasActiveCharacter()); // code in this function couldn't handle pParty->activeCharacterIndex() = 0 and crash
-
-    int pPrice = PriceCalculator::transportCostForPlayer(&pParty->activeCharacter(),
-                                                         buildingTable[window_SpeakInHouse->wData.val - 1]);
-    int route_id = window_SpeakInHouse->wData.val - HOUSE_STABLES_HARMONDALE;
-
-    if (dialog_menu_id == DIALOGUE_MAIN) {
-        if (HouseUI_CheckIfPlayerCanInteract()) {
-            int index = 0;
-
-            std::string travelcost = localization->FormatString(LSTR_FMT_TRAVEL_COST_D_GOLD, pPrice);
-            int pTextHeight = pFontArrus->CalcTextHeight(travelcost, travel_window.uFrameWidth, 0);
-            int pRealTextHeight = pTextHeight + (pFontArrus->GetHeight() - 3) + 146;
-            int pPrimaryTextHeight = pRealTextHeight;
-
-            int pCurrentButton = 2;
-            int lastsched{ 255 };
-            std::string pTopicArray[5]{};
-
-            for (uint i = pDialogueWindow->pStartingPosActiveItem; i < (unsigned int)(pDialogueWindow->pNumPresenceButton + pDialogueWindow->pStartingPosActiveItem); ++i) {
-                int schedule_id = transport_routes[route_id][index];
-                GUIButton *pButton = pDialogueWindow->GetControl(i);
-
-                bool route_active = 0;
-
-                if (schedule_id != 255 && (lastsched != schedule_id)) {
-                    Assert(schedule_id < 35);
-                    if (pCurrentButton >= 6)
-                        route_active = true;
-                    else
-                        route_active = transport_schedule[schedule_id].pSchedule[pParty->uCurrentDayOfMonth % 7];
-                }
-
-                lastsched = schedule_id;
-
-                if (schedule_id != 255 && route_active &&
-                    (!transport_schedule[schedule_id].uQuestBit || pParty->_questBits[transport_schedule[schedule_id].uQuestBit])) {
-                    uint16_t color{};
-                    if (pDialogueWindow->pCurrentPosActiveItem == pCurrentButton)
-                        color = colorTable.PaleCanary.c16();
-                    else
-                        color = colorTable.White.c16();
-
-                    pTopicArray[index] = fmt::format("\f{:05}", color);
-
-                    int travel_time = GetTravelTimeTransportDays(schedule_id);
-
-                    std::string str = localization->FormatString(LSTR_FMT_D_DAYS_TO_S, travel_time, pMapStats->pInfos[transport_schedule[schedule_id].uMapInfoID].pName.c_str());
-                    pTopicArray[index] += str + "\n \n";
-                    pButton->uY = pRealTextHeight;
-                    pTextHeight = pFontArrus->CalcTextHeight(str, travel_window.uFrameWidth, 0);
-                    pButton->uHeight = pTextHeight;
-                    pButton->uW = pButton->uY + pTextHeight - 1 + 6;
-                    pRealTextHeight += (pFontArrus->GetHeight() - 3) + pTextHeight;
-                } else {
-                    pTopicArray[index] = "";
-                    if (pButton) {
-                        pButton->uW = 0;
-                        pButton->uHeight = 0;
-                        pButton->uY = 0;
-                    }
-                }
-                ++index;
-                ++pCurrentButton;
-            }
-
-            if (pRealTextHeight != pPrimaryTextHeight) {
-                // height differences means we have travel options
-                travel_window.DrawTitleText(pFontArrus, 0, 146, 0,
-                    fmt::format("{}\n \n{}{}{}{}{}", travelcost, pTopicArray[0], pTopicArray[1], pTopicArray[2], pTopicArray[3], pTopicArray[4]), 3);
-            } else {
-                travel_window.DrawTitleText(pFontArrus, 0, (SIDE_TEXT_BOX_BODY_TEXT_HEIGHT - pFontArrus->CalcTextHeight(localization->GetString(LSTR_COME_BACK_ANOTHER_DAY), travel_window.uFrameWidth, 0)) / 2 + SIDE_TEXT_BOX_BODY_TEXT_OFFSET,
-                                            colorTable.White.c16(), localization->GetString(LSTR_COME_BACK_ANOTHER_DAY), 3);
-            }
-        }
-    } else {  //после нажатия топика - travel option selected
-        if (dialog_menu_id >= DIALOGUE_TRANSPORT_SCHEDULE_1 &&
-            dialog_menu_id <= DIALOGUE_TRANSPORT_SCHEDULE_4) {
-            if (pParty->GetGold() < pPrice) {
-                GameUI_SetStatusBar(LSTR_NOT_ENOUGH_GOLD);
-                // TODO(pskelton): correct sound but wrong label - travel house sounds might need different enum
-                PlayHouseSound(window_SpeakInHouse->wData.val, HouseSound_Greeting_2);
-                pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-                return;
-            }
-
-            int choice_id = dialog_menu_id - DIALOGUE_TRANSPORT_SCHEDULE_1;
-            stru365_travel_info *pTravel = &transport_schedule[transport_routes[route_id][choice_id]];
-
-            if (pTravel->pSchedule[pParty->uCurrentDayOfMonth % 7]) {
-                if (pCurrentMapName != pMapStats->pInfos[pTravel->uMapInfoID].pFilename) {
-                    SaveGame(1, 0);
-                    pCurrentMapName = pMapStats->pInfos[pTravel->uMapInfoID].pFilename;
-
-                    dword_6BE364_game_settings_1 |= GAME_SETTINGS_0001;
-                    Party_Teleport_Cam_Pitch = 0;
-                    Party_Teleport_Z_Speed = 0;
-                    Party_Teleport_Cam_Yaw = pTravel->arrival_view_yaw;
-                    uGameState = GAME_STATE_CHANGE_LOCATION;
-                    Party_Teleport_X_Pos = pTravel->arrival_x;
-                    Party_Teleport_Y_Pos = pTravel->arrival_y;
-                    Party_Teleport_Z_Pos = pTravel->arrival_z;
-                    Start_Party_Teleport_Flag = pTravel->arrival_x | pTravel->arrival_y | pTravel->arrival_z | pTravel->arrival_view_yaw;
-                } else {
-                    // travelling to map we are already in
-                    pCamera3D->_viewYaw = 0;
-
-                    pParty->uFlags |= PARTY_FLAGS_1_ForceRedraw;
-                    pParty->vPosition.x = pTravel->arrival_x;
-                    pParty->vPosition.y = pTravel->arrival_y;
-                    pParty->vPosition.z = pTravel->arrival_z;
-                    pParty->uFallStartZ = pParty->vPosition.z;
-                    pParty->_viewPitch = 0;
-                    pParty->_viewYaw = pTravel->arrival_view_yaw;
-                }
-
-                pParty->TakeGold(pPrice);
-                // TODO(pskelton): correct sound but wrong label - travel house sounds might need different enum
-                PlayHouseSound(window_SpeakInHouse->wData.val, HouseSound_NotEnoughMoney);
-                int traveltimedays = GetTravelTimeTransportDays(transport_routes[route_id][choice_id]);
-
-                PlayerSpeech pSpeech;
-                if (route_id >= HOUSE_BOATS_EMERALD_ISLE - HOUSE_STABLES_HARMONDALE) {
-                    pSpeech = SPEECH_TravelBoat;
-                } else {
-                    pSpeech = SPEECH_TravelHorse;
-                }
-
-                restAndHeal(GameTime::FromDays(traveltimedays));
-                pParty->activeCharacter().playReaction(pSpeech);
-                pAudioPlayer->soundDrain();
-                while (HouseDialogPressCloseBtn()) {}
-                pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 0, 0);
-            } else {
-                dialog_menu_id = DIALOGUE_MAIN;
-                pAudioPlayer->playUISound(SOUND_error);
-            }
-        }
-    }
-}
-
-bool IsTravelAvailable(int a1) {
-    for (uint i = 0; i < 4; ++i) {
-        if (transport_schedule[transport_routes[a1][i]]
-            .pSchedule[pParty->uCurrentDayOfMonth % 7]) {
-            if (!transport_schedule[transport_routes[a1][i]].uQuestBit || pParty->_questBits[transport_schedule[transport_routes[a1][i]].uQuestBit]) {
-                return true;
-            }
-        }
-    }
-    return false;
-}
-
 //----- (004B7911) --------------------------------------------------------
 void TownHallDialog() {
     int v1;                       // eax@10
@@ -1520,895 +1266,6 @@ void TownHallDialog() {
         break;
     }
     return;
-}
-
-void BankDialog() {
-    GUIWindow bank_window = *window_SpeakInHouse;
-    bank_window.uFrameX = SIDE_TEXT_BOX_POS_X;
-    bank_window.uFrameWidth = SIDE_TEXT_BOX_WIDTH;
-    bank_window.uFrameZ = SIDE_TEXT_BOX_POS_Z;
-    bank_window.DrawTitleText(
-        pFontArrus, 0, 220, colorTable.PaleCanary.c16(),
-        fmt::format("{}: {}", localization->GetString(LSTR_BALANCE), pParty->uNumGoldInBank),
-        3);
-    switch (dialog_menu_id) {
-    case DIALOGUE_MAIN:
-    {
-        uint16_t pColorText = colorTable.PaleCanary.c16();
-        if (pDialogueWindow->pCurrentPosActiveItem != 2) {
-            pColorText = colorTable.White.c16();
-        }
-        bank_window.DrawTitleText(pFontArrus, 0, 146, pColorText,
-            localization->GetString(LSTR_DEPOSIT), 3);
-        pColorText = colorTable.PaleCanary.c16();
-        if (pDialogueWindow->pCurrentPosActiveItem != 3) {
-            pColorText = colorTable.White.c16();
-        }
-        bank_window.DrawTitleText(pFontArrus, 0, 176, pColorText,
-            localization->GetString(LSTR_WITHDRAW), 3);
-        break;
-    }
-    case DIALOGUE_BANK_PUT_GOLD:
-    {
-        if (window_SpeakInHouse->keyboard_input_status == WINDOW_INPUT_IN_PROGRESS) {
-            bank_window.DrawTitleText(pFontArrus, 0, 146, colorTable.PaleCanary.c16(),
-                                      fmt::format("{}\n{}", localization->GetString(LSTR_DEPOSIT), localization->GetString(LSTR_HOW_MUCH)), 3);
-            bank_window.DrawTitleText(pFontArrus, 0, 186, colorTable.White.c16(), keyboardInputHandler->GetTextInput().c_str(), 3);
-            bank_window.DrawFlashingInputCursor(pFontArrus->GetLineWidth(keyboardInputHandler->GetTextInput().c_str()) / 2 + 80, 185, pFontArrus);
-            return;
-        }
-        if (window_SpeakInHouse->keyboard_input_status == WINDOW_INPUT_CONFIRMED) {
-            int sum = atoi(keyboardInputHandler->GetTextInput().c_str());
-            if (sum <= 0) {
-                pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-                return;
-            }
-
-            int party_gold = pParty->GetGold();
-            if (sum > party_gold) {
-                PlayHouseSound(window_SpeakInHouse->wData.val, HouseSound_NotEnoughMoney);
-                sum = party_gold;
-            }
-
-            if (sum > 0) {
-                pParty->TakeGold(sum);
-                pParty->AddBankGold(sum);
-                if (pParty->hasActiveCharacter()) {
-                    pParty->activeCharacter().playReaction(SPEECH_BankDeposit);
-                }
-            }
-        }
-        window_SpeakInHouse->keyboard_input_status = WINDOW_INPUT_NONE;
-        pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-        break;
-    }
-    case DIALOGUE_BANK_GET_GOLD:
-    {
-        if (window_SpeakInHouse->keyboard_input_status == WINDOW_INPUT_IN_PROGRESS) {
-            bank_window.DrawTitleText(pFontArrus, 0, 146, colorTable.PaleCanary.c16(),
-                                      fmt::format("{}\n{}", localization->GetString(LSTR_WITHDRAW), localization->GetString(LSTR_HOW_MUCH)), 3);
-            bank_window.DrawTitleText(pFontArrus, 0, 186, colorTable.White.c16(), keyboardInputHandler->GetTextInput().c_str(), 3);
-            bank_window.DrawFlashingInputCursor(pFontArrus->GetLineWidth(keyboardInputHandler->GetTextInput().c_str()) / 2 + 80, 185, pFontArrus);
-            return;
-        } else if (window_SpeakInHouse->keyboard_input_status == WINDOW_INPUT_CONFIRMED) {
-            window_SpeakInHouse->keyboard_input_status = WINDOW_INPUT_NONE;
-            int sum = atoi(keyboardInputHandler->GetTextInput().c_str());
-            if (sum <= 0) {
-                pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-                return;
-            }
-
-            int bank_gold = pParty->GetBankGold();
-            if (sum > bank_gold) {
-                PlayHouseSound(window_SpeakInHouse->wData.val, HouseSound_NotEnoughMoney);
-                sum = bank_gold;
-            }
-
-            if (sum > 0) {
-                pParty->TakeBankGold(sum);
-                pParty->AddGold(sum);
-            }
-        }
-        window_SpeakInHouse->keyboard_input_status = WINDOW_INPUT_NONE;
-        pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-        break;
-    }
-    default:
-        break;
-    }
-}
-
-//----- (004B8285) --------------------------------------------------------
-void TavernDialog() {
-    int pItemNum;
-    double v2;                // st7@1
-    int pNumString;           // edi@16
-    GUIButton *pButton;       // eax@65
-    int pSkillCount;
-    signed int pOptionsCount;       // edi@77
-    signed int i;                   // esi@79
-                                    //  signed int v53; // edi@81
-    const char *pText;              // [sp-4h] [bp-278h]@93
-    unsigned int pTopic1Height;     // [sp+26Fh] [bp-5h]@55
-    uint8_t pTopic2Height;  // [sp+267h] [bp-Dh]@57
-    uint8_t pTopic3Height;  // [sp+253h] [bp-21h]@59
-    uint8_t pTopic4Height = 0;
-    int pTextHeight;
-    int all_text_height = 0;  // [sp+260h] [bp-14h]@18
-    GUIFont *pOutString;
-
-    GUIWindow dialog_window = *window_SpeakInHouse;
-    dialog_window.uFrameX = SIDE_TEXT_BOX_POS_X;
-    dialog_window.uFrameWidth = SIDE_TEXT_BOX_WIDTH;
-    dialog_window.uFrameZ = SIDE_TEXT_BOX_POS_Z;
-
-    // TODO(pskelton): check this behaviour
-    if (!pParty->hasActiveCharacter())  // avoid nzi
-        pParty->setActiveToFirstCanAct();
-
-    const BuildingDesc &house = buildingTable[window_SpeakInHouse->wData.val - 1];
-
-    int pPriceRoom = PriceCalculator::tavernRoomCostForPlayer(&pParty->activeCharacter(), house);
-    int pPriceFood = PriceCalculator::tavernFoodCostForPlayer(&pParty->activeCharacter(), house);
-
-    switch (dialog_menu_id) {
-    case DIALOGUE_MAIN:
-    {
-        if (!HouseUI_CheckIfPlayerCanInteract()) return;
-
-        std::string topic1 = fmt::format("\f{:05}",
-                pDialogueWindow->pCurrentPosActiveItem == 2
-                ? colorTable.PaleCanary.c16()
-                : colorTable.White.c16()) +
-                             localization->FormatString(LSTR_FMT_RENT_ROOM_FOR_D_GOLD, pPriceRoom);
-        pTopic1Height = pFontArrus->CalcTextHeight(
-            topic1, dialog_window.uFrameWidth, 0);
-
-        std::string topic2 = fmt::format("\f{:05}",
-                pDialogueWindow->pCurrentPosActiveItem == 3
-                ? colorTable.PaleCanary.c16()
-                : colorTable.White.c16())
-            + localization->FormatString(
-                LSTR_FMT_BUY_D_FOOD_FOR_D_GOLD,
-                (unsigned int)
-                buildingTable[window_SpeakInHouse->wData.val - 1].fPriceMultiplier,
-                pPriceFood);
-        pTopic2Height = pFontArrus->CalcTextHeight(topic2, dialog_window.uFrameWidth, 0);
-
-        std::string topic3 = fmt::format("\f{:05}",
-                pDialogueWindow->pCurrentPosActiveItem == 4
-                ? colorTable.PaleCanary.c16()
-                : colorTable.White.c16()) +
-                             localization->GetString(LSTR_LEARN_SKILLS);
-        pTopic3Height = pFontArrus->CalcTextHeight(topic3, dialog_window.uFrameWidth, 0);
-
-        std::string topic4;
-        if (window_SpeakInHouse->wData.val >= 108 &&
-            window_SpeakInHouse->wData.val <= 120) {
-            topic4 =
-                fmt::format("\f{:05}",
-                    pDialogueWindow->pCurrentPosActiveItem == 5
-                    ? colorTable.PaleCanary.c16()
-                    : colorTable.White.c16()) +
-                localization->GetString(LSTR_PLAY_ARCOMAGE);
-            pTopic4Height = pFontArrus->CalcTextHeight(
-                topic4, dialog_window.uFrameWidth, 0);
-        }
-
-        if (pDialogueWindow->pNumPresenceButton) {
-            for (pItemNum = pDialogueWindow->pStartingPosActiveItem;
-                pItemNum < pDialogueWindow->pStartingPosActiveItem +
-                pDialogueWindow->pNumPresenceButton;
-                ++pItemNum) {
-                pButton = pDialogueWindow->GetControl(pItemNum);
-                if (pButton->msg_param == 15) {
-                    pButton->uHeight = pTopic1Height;
-                    pButton->uY = 146;
-                    pButton->uW = pTopic1Height + 145 + 6;
-                } else if (pButton->msg_param == 16) {
-                    pButton->uHeight = pTopic2Height;
-                    pButton->uY =
-                        pTopic1Height + (pFontArrus->GetHeight() ) + 146;
-                    pButton->uW = (pTopic1Height +
-                        (pFontArrus->GetHeight() ) + 146) +
-                        pTopic2Height - 1 + 6;
-                } else if (pButton->msg_param == 96) {
-                    pButton->uY = pTopic1Height + pTopic2Height +
-                        2 * (pFontArrus->GetHeight() ) + 146;
-                    pButton->uHeight = pTopic3Height;
-                    pButton->uW =
-                        pTopic3Height +
-                        (pTopic1Height + pTopic2Height +
-                            2 * (pFontArrus->GetHeight() ) + 146) -
-                        1 + 6;
-                } else if (pButton->msg_param == 101) {
-                    pButton->uHeight = pTopic4Height;
-                    pButton->uY = pTopic1Height +
-                        3 * (pFontArrus->GetHeight() ) +
-                        pTopic4Height + pTopic2Height + 146;
-                    pButton->uW =
-                        (pTopic1Height + 3 * (pFontArrus->GetHeight() ) +
-                            pTopic4Height + pTopic2Height + 146) +
-                        pTopic4Height - 1 + 6;
-                }
-            }
-            dialog_window.DrawTitleText(
-                pFontArrus, 0, 146, 0,
-                fmt::format("{}\n \n{}\n \n{}\n \n{}", topic1, topic2, topic3, topic4),
-                3);
-        }
-        break;
-    }
-    case DIALOGUE_TAVERN_ARCOMAGE_RULES:
-    {
-        pOutString = pFontArrus;
-
-        std::string str = pNPCTopics[354].pText;
-        dialog_window.uFrameWidth = game_viewport_width;
-        dialog_window.uFrameZ = 452;
-        pTextHeight = pFontArrus->CalcTextHeight(str, dialog_window.uFrameWidth, 12) + 7;
-        if (352 - pTextHeight < 8) {
-            pOutString = pFontCreate;
-            pTextHeight = pFontCreate->CalcTextHeight(str, dialog_window.uFrameWidth, 12) + 7;
-        }
-        render->DrawTextureCustomHeight(8 / 640.0f, (352 - pTextHeight) / 480.0f, ui_leather_mm7, pTextHeight);
-        render->DrawTextureNew(8 / 640.0f, (347 - pTextHeight) / 480.0f, _591428_endcap);
-        window_SpeakInHouse->DrawText(pOutString, {12, 354 - pTextHeight}, colorTable.Black.c16(),
-            pOutString->FitTextInAWindow(str, dialog_window.uFrameWidth, 12), 0, 0, 0);
-        break;
-    }
-
-    case DIALOGUE_TAVERN_ARCOMAGE_VICTORY_CONDITIONS:
-    {
-        std::string label =
-            pNPCTopics[window_SpeakInHouse->wData.val + 247].pText;
-        dialog_window.uFrameWidth = game_viewport_width;
-        dialog_window.uFrameZ = 452;
-        pTextHeight = pFontArrus->CalcTextHeight(
-            label, dialog_window.uFrameWidth, 12) +
-            7;
-        render->DrawTextureCustomHeight(8 / 640.0f,
-            (352 - pTextHeight) / 480.0f,
-            ui_leather_mm7, pTextHeight);
-        render->DrawTextureNew(
-            8 / 640.0f, (347 - pTextHeight) / 480.0f, _591428_endcap);
-        window_SpeakInHouse->DrawText(pFontArrus, {12, 354 - pTextHeight}, colorTable.Black.c16(),
-            pFontArrus->FitTextInAWindow(label, dialog_window.uFrameWidth, 12));
-        break;
-    }
-
-    case DIALOGUE_TAVERN_ARCOMAGE_RESULT:
-    {
-        if (pArcomageGame->bGameInProgress == 1) return;
-        if (pArcomageGame->uGameWinner) {
-            if (pArcomageGame->uGameWinner == 1)
-                pText = localization->GetString(LSTR_YOU_WON);
-            else
-                pText = localization->GetString(LSTR_YOU_LOST);
-        } else {
-            pText = localization->GetString(LSTR_A_TIE);
-        }
-        dialog_window.DrawTitleText(
-            pFontArrus, 0,
-            (SIDE_TEXT_BOX_BODY_TEXT_HEIGHT - pFontArrus->CalcTextHeight(
-                pText, dialog_window.uFrameWidth, 0)) /
-            2 +
-            SIDE_TEXT_BOX_BODY_TEXT_OFFSET, colorTable.PaleCanary.c16(), pText, 3);
-        break;
-    }
-    case DIALOGUE_TAVERN_REST:
-    {
-        if (pParty->GetGold() >= pPriceRoom) {
-            pParty->TakeGold(pPriceRoom);
-            PlayHouseSound(window_SpeakInHouse->wData.val,
-                HouseSound_NotEnoughMoney);
-            dialog_menu_id = DIALOGUE_NULL;
-            HouseDialogPressCloseBtn();
-            GetHouseGoodbyeSpeech();
-            pMediaPlayer->Unload();
-
-            pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_RentRoom, window_SpeakInHouse->wData.val, 1);
-            window_SpeakInHouse->Release();
-            window_SpeakInHouse = 0;
-            return;
-        }
-        GameUI_SetStatusBar(LSTR_NOT_ENOUGH_GOLD);
-        PlayHouseSound(window_SpeakInHouse->wData.val, HouseSound_Goodbye);
-        pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-        break;
-    }
-
-    case DIALOGUE_LEARN_SKILLS:
-    {
-        if (!HouseUI_CheckIfPlayerCanInteract()) return;
-        pSkillCount = 0;
-        int pPriceSkill = PriceCalculator::skillLearningCostForPlayer(&pParty->activeCharacter(), house);
-        all_text_height = 0;
-        for (int i = pDialogueWindow->pStartingPosActiveItem;
-            i < pDialogueWindow->pStartingPosActiveItem +
-            pDialogueWindow->pNumPresenceButton;
-            ++i) {
-            auto skill = GetLearningDialogueSkill(
-                (DIALOGUE_TYPE)pDialogueWindow->GetControl(i)->msg_param
-            );
-            if (skillMaxMasteryPerClass[pParty->activeCharacter().classType][skill] != PLAYER_SKILL_MASTERY_NONE
-                && !pParty->activeCharacter().pActiveSkills[skill]) {
-                all_text_height = pFontArrus->CalcTextHeight(
-                    localization->GetSkillName(skill),
-                    dialog_window.uFrameWidth, 0);
-                pSkillCount++;
-            }
-        }
-        SkillTrainingDialogue(&dialog_window, pSkillCount, all_text_height, pPriceSkill);
-        return;
-    }
-
-    case DIALOGUE_TAVERN_BUY_FOOD:
-    {
-        if ((double)pParty->GetFood() >=
-            buildingTable[window_SpeakInHouse->wData.val - 1].fPriceMultiplier) {
-            GameUI_SetStatusBar(LSTR_RATIONS_FULL);
-            if (pParty->hasActiveCharacter())
-                pParty->activeCharacter().playReaction(SPEECH_PacksFull);
-            pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-            return;
-        }
-        if (pParty->GetGold() >= pPriceFood) {
-            pParty->TakeGold(pPriceFood);
-            pParty->SetFood(buildingTable[window_SpeakInHouse->wData.val - 1].fPriceMultiplier);
-            PlayHouseSound(window_SpeakInHouse->wData.val, HouseSound_Greeting_2);
-            pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-            return;
-        }
-        GameUI_SetStatusBar(LSTR_NOT_ENOUGH_GOLD);
-        PlayHouseSound(window_SpeakInHouse->wData.val, HouseSound_Goodbye);
-        pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-        break;
-    }
-
-    case DIALOGUE_TAVERN_ARCOMAGE_MAIN:
-    {
-        if (HouseUI_CheckIfPlayerCanInteract()) {
-            pOptionsCount = 2;
-            pShopOptions[0] = localization->GetString(LSTR_RULES);
-            pShopOptions[1] = localization->GetString(LSTR_VICTORY_CONDITIONS);
-            if (pParty->hasItem(ITEM_QUEST_ARCOMAGE_DECK)) {
-                pShopOptions[2] = localization->GetString(LSTR_PLAY);
-                pOptionsCount = 3;
-            }
-            for (i = 0; i < pOptionsCount; ++i)
-                all_text_height += pFontArrus->CalcTextHeight(pShopOptions[i], dialog_window.uFrameWidth, 0);
-            all_text_height = (SIDE_TEXT_BOX_BODY_TEXT_HEIGHT - all_text_height) / pOptionsCount;
-
-            int v54 = (SIDE_TEXT_BOX_BODY_TEXT_HEIGHT - pOptionsCount * all_text_height - all_text_height) /
-                2 - all_text_height / 2 + SIDE_TEXT_BOX_BODY_TEXT_OFFSET;
-            pNumString = 0;
-            for (pItemNum = pDialogueWindow->pStartingPosActiveItem;
-                pItemNum < pDialogueWindow->pNumPresenceButton +
-                pDialogueWindow->pStartingPosActiveItem;
-                pItemNum++) {
-                pButton = pDialogueWindow->GetControl(pItemNum);
-                pButton->uY = all_text_height + v54;
-                pTextHeight = pFontArrus->CalcTextHeight(
-                    pShopOptions[pNumString], dialog_window.uFrameWidth, 0);
-                pButton->uHeight = pTextHeight;
-                v54 = pButton->uY + pTextHeight - 1;
-                pButton->uW = v54 + 6;
-                int pColorText = colorTable.PaleCanary.c16();
-                if (pDialogueWindow->pCurrentPosActiveItem != pItemNum)
-                    pColorText = colorTable.White.c16();
-                dialog_window.DrawTitleText(pFontArrus, 0, pButton->uY,
-                    pColorText,
-                    pShopOptions[pNumString], 3);
-                ++pNumString;
-            }
-        }
-        break;
-    }
-    default: { break; }
-    }
-}
-
-void TempleDialog() {
-    int pPrice;                   // edi@1
-    int pTextHeight;              // eax@11
-    uint16_t pTextColor;  // ax@21
-    LocationInfo *ddm;          // edi@29
-    GUIButton *pButton;           // edi@64
-    uint8_t index;        // [sp+1B7h] [bp-Dh]@64
-    int v64;                      // [sp+1B8h] [bp-Ch]@6
-    unsigned int pCurrentItem;    // [sp+1BCh] [bp-8h]@6
-    int all_text_height;          // [sp+1C0h] [bp-4h]@6
-
-    GUIWindow tample_window = *window_SpeakInHouse;
-    tample_window.uFrameX = SIDE_TEXT_BOX_POS_X;
-    tample_window.uFrameWidth = SIDE_TEXT_BOX_WIDTH;
-    tample_window.uFrameZ = SIDE_TEXT_BOX_POS_Z;
-
-    // TODO(pskelton): check this behaviour
-    if (!pParty->hasActiveCharacter()) {  // avoid nzi
-        pParty->setActiveToFirstCanAct();
-    }
-
-    pPrice = PriceCalculator::templeHealingCostForPlayer(&pParty->activeCharacter(),
-                                                         buildingTable[window_SpeakInHouse->wData.val - 1].fPriceMultiplier);
-    if (dialog_menu_id == DIALOGUE_MAIN) {
-        index = 1;
-        pButton = pDialogueWindow->GetControl(
-            pDialogueWindow->pStartingPosActiveItem);
-        pButton->uHeight = 0;
-        pButton->uY = 0;
-        if (pParty->activeCharacter().IsPlayerHealableByTemple()) {
-            static std::string shop_option_container;
-            shop_option_container =
-                fmt::format("{} {} {}",
-                    localization->GetString(LSTR_HEAL), pPrice,
-                    localization->GetString(LSTR_GOLD));
-            pShopOptions[0] = shop_option_container.c_str();
-            index = 0;
-        }
-        pShopOptions[1] = localization->GetString(LSTR_DONATE);
-        pShopOptions[2] = localization->GetString(LSTR_LEARN_SKILLS);
-        all_text_height = 0;
-        if (index < pDialogueWindow->pNumPresenceButton) {
-            uint i = index;
-            for (uint j = index; j < pDialogueWindow->pNumPresenceButton; ++j) {
-                all_text_height += pFontArrus->CalcTextHeight(
-                    pShopOptions[1 * i], tample_window.uFrameWidth, 0);
-                i++;
-            }
-        }
-        v64 = (SIDE_TEXT_BOX_BODY_TEXT_HEIGHT - (signed int)all_text_height) /
-            (pDialogueWindow->pNumPresenceButton - index);
-        if (v64 > SIDE_TEXT_BOX_MAX_SPACING) v64 = SIDE_TEXT_BOX_MAX_SPACING;
-        all_text_height =
-            (SIDE_TEXT_BOX_BODY_TEXT_HEIGHT - v64 * (pDialogueWindow->pNumPresenceButton - index) -
-            (signed int)all_text_height) /
-            2 -
-            v64 / 2 + SIDE_TEXT_BOX_BODY_TEXT_OFFSET;
-        if (index + pDialogueWindow->pStartingPosActiveItem <
-            pDialogueWindow->pStartingPosActiveItem +
-            pDialogueWindow->pNumPresenceButton) {
-            uint i = index;
-            for (pCurrentItem = index + pDialogueWindow->pStartingPosActiveItem;
-                (signed int)pCurrentItem <
-                pDialogueWindow->pNumPresenceButton +
-                pDialogueWindow->pStartingPosActiveItem;
-                ++pCurrentItem) {
-                pButton = pDialogueWindow->GetControl(pCurrentItem);
-                pButton->uY = v64 + all_text_height;
-                pTextHeight = pFontArrus->CalcTextHeight(
-                    pShopOptions[1 * i], tample_window.uFrameWidth, 0);
-                pButton->uHeight = pTextHeight;
-                pButton->uW = pButton->uY + pTextHeight - 1 + 6;
-                all_text_height = pButton->uW;
-                pTextColor = colorTable.PaleCanary.c16();
-                if (pDialogueWindow->pCurrentPosActiveItem != index + 2)
-                    pTextColor = colorTable.White.c16();
-                tample_window.DrawTitleText(pFontArrus, 0, pButton->uY,
-                    pTextColor, pShopOptions[1 * i], 3);
-                i++;
-                index++;
-            }
-        }
-        return;
-    }
-    //-------------------------------------------------
-    if (dialog_menu_id == DIALOGUE_TEMPLE_HEAL) {
-        if (!pParty->activeCharacter().IsPlayerHealableByTemple()) return;
-        if (pParty->GetGold() < pPrice) {
-            GameUI_SetStatusBar(LSTR_NOT_ENOUGH_GOLD);
-            PlayHouseSound(window_SpeakInHouse->wData.val, HouseSound_NotEnoughMoney);
-            pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-            return;
-        }
-        pParty->TakeGold(pPrice);
-
-        pParty->activeCharacter().conditions.ResetAll();
-        pParty->activeCharacter().health =
-            pParty->activeCharacter().GetMaxHealth();
-        pParty->activeCharacter().mana =
-            pParty->activeCharacter().GetMaxMana();
-
-        if (window_SpeakInHouse->wData.val != 78 &&
-            (window_SpeakInHouse->wData.val <= 80 ||
-            window_SpeakInHouse->wData.val > 82)) {
-            if (pParty->activeCharacter().conditions.Has(Condition_Zombie)) {
-                // если состояние зомби
-                pParty->activeCharacter().uCurrentFace =
-                    pParty->activeCharacter().uPrevFace;
-                pParty->activeCharacter().uVoiceID =
-                    pParty->activeCharacter().uPrevVoiceID;
-                GameUI_ReloadPlayerPortraits(
-                    pParty->activeCharacterIndex() - 1,
-                    pParty->activeCharacter().uPrevFace);
-            }
-            pAudioPlayer->playExclusiveSound(SOUND_heal);
-            pParty->activeCharacter().playReaction(SPEECH_TempleHeal);
-            pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-            return;
-        }
-        if (pParty->activeCharacter().conditions.Has(Condition_Zombie)) {
-            // LODWORD(pParty->activeCharacter().pConditions[Condition_Zombie])
-            // =
-            // LODWORD(pParty->activeCharacter().pConditions[Condition_Zombie]);
-        } else {
-            if (pParty->activeCharacter().conditions.HasNone({Condition_Eradicated, Condition_Petrified, Condition_Dead})) {
-                pAudioPlayer->playExclusiveSound(SOUND_heal);
-                pParty->activeCharacter().playReaction(SPEECH_TempleHeal);
-                pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-                return;
-            }
-            pParty->activeCharacter().uPrevFace =
-                pParty->activeCharacter().uCurrentFace;
-            pParty->activeCharacter().uPrevVoiceID =
-                pParty->activeCharacter().uVoiceID;
-            pParty->activeCharacter().SetCondition(Condition_Zombie, 1);
-            pParty->activeCharacter().uVoiceID =
-                (pParty->activeCharacter().GetSexByVoice() != 0) + 23;
-            pParty->activeCharacter().uCurrentFace =
-                (pParty->activeCharacter().GetSexByVoice() != 0) + 23;
-            GameUI_ReloadPlayerPortraits(
-                pParty->activeCharacterIndex() - 1,
-                (pParty->activeCharacter().GetSexByVoice() != 0) + 23);
-            pParty->activeCharacter().conditions.Set(Condition_Zombie, pParty->GetPlayingTime());
-            // v39 = (GUIWindow *)HIDWORD(pParty->uTimePlayed);
-        }
-        // HIDWORD(pParty->activeCharacter().pConditions[Condition_Zombie]) =
-        // (int)v39;
-        pParty->activeCharacter().conditions.Set(Condition_Zombie, pParty->GetPlayingTime());
-        pAudioPlayer->playExclusiveSound(SOUND_heal);
-        pParty->activeCharacter().playReaction(SPEECH_TempleHeal);
-        pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-        return;
-    }
-    //---------------------------------------------------
-    if (dialog_menu_id == DIALOGUE_TEMPLE_DONATE) {
-        pPrice = buildingTable[window_SpeakInHouse->wData.val - 1].fPriceMultiplier;
-        if (pParty->GetGold() >= pPrice) {
-            pParty->TakeGold(pPrice);
-            ddm = &currentLocationInfo();
-
-            if (ddm->reputation > -5) {
-                ddm->reputation = ddm->reputation - 1;
-                if (ddm->reputation - 1 < -5) ddm->reputation = -5;
-            }
-            if ((uint8_t)byte_F8B1EF[pParty->activeCharacterIndex()] == pParty->uCurrentDayOfMonth % 7) {
-                if (ddm->reputation <= -5) {
-                    pushTempleSpell(SPELL_AIR_WIZARD_EYE);
-                }
-                if (ddm->reputation <= -10) {
-                    pushTempleSpell(SPELL_SPIRIT_PRESERVATION);
-                }
-                if (ddm->reputation <= -15) {
-                    pushTempleSpell(SPELL_BODY_PROTECTION_FROM_MAGIC);
-                }
-                if (ddm->reputation <= -20) {
-                    pushTempleSpell(SPELL_LIGHT_HOUR_OF_POWER);
-                }
-                if (ddm->reputation <= -25) {
-                    pushTempleSpell(SPELL_LIGHT_DAY_OF_PROTECTION);
-                }
-            }
-            ++byte_F8B1EF[pParty->activeCharacterIndex()];
-            pParty->activeCharacter().playReaction(SPEECH_TempleDonate);
-            GameUI_SetStatusBar(LSTR_THANK_YOU);
-            pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-            return;
-        }
-        GameUI_SetStatusBar(LSTR_NOT_ENOUGH_GOLD);
-        PlayHouseSound(window_SpeakInHouse->wData.val,
-            HouseSound_NotEnoughMoney);
-        pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-        return;
-    }
-    //------------------------------------------------
-    if (dialog_menu_id == DIALOGUE_LEARN_SKILLS) {
-        if (HouseUI_CheckIfPlayerCanInteract()) {
-            all_text_height = 0;
-            v64 = PriceCalculator::skillLearningCostForPlayer(&pParty->activeCharacter(),
-                                                              buildingTable[window_SpeakInHouse->wData.val - 1]);
-            pCurrentItem = 0;
-            for (int i = pDialogueWindow->pStartingPosActiveItem;
-                i < pDialogueWindow->pNumPresenceButton +
-                pDialogueWindow->pStartingPosActiveItem;
-                ++i) {
-                auto skill = GetLearningDialogueSkill(
-                    (DIALOGUE_TYPE)pDialogueWindow->GetControl(i)->msg_param
-                );
-                if (skillMaxMasteryPerClass[pParty->activeCharacter().classType][skill] != PLAYER_SKILL_MASTERY_NONE
-                    && !pParty->activeCharacter().pActiveSkills[skill]) {
-                    all_text_height += pFontArrus->CalcTextHeight(
-                        localization->GetSkillName(skill),
-                        tample_window.uFrameWidth, 0);
-                    ++pCurrentItem;
-                }
-            }
-
-            SkillTrainingDialogue(&tample_window, pCurrentItem, all_text_height, v64);
-        }
-    }
-}
-
-void TrainingDialog(const char *s) {
-    uint64_t expForNextLevel;  // edi@3
-    int v8;               // edx@4
-    double v9;            // st7@6
-    signed int v10;       // esi@6
-    int v14;              // esi@14
-    int v33;              // eax@36
-    unsigned int v36;     // eax@38
-    int index;
-    int all_text_height;          // eax@68
-    int v49;                      // ebx@69
-    GUIButton *pButton;           // eax@71
-    int pTextHeight;              // eax@71
-
-    GUIWindow training_dialog_window = *window_SpeakInHouse;
-    training_dialog_window.uFrameX = SIDE_TEXT_BOX_POS_X;
-    training_dialog_window.uFrameWidth = SIDE_TEXT_BOX_WIDTH;
-    training_dialog_window.uFrameZ = SIDE_TEXT_BOX_POS_Z;
-
-    // TODO(pskelton): check this behaviour
-    if (!pParty->hasActiveCharacter())  // avoid nzi
-        pParty->setActiveToFirstCanAct();
-
-    int pPrice = PriceCalculator::trainingCostForPlayer(&pParty->activeCharacter(),
-                                                        buildingTable[window_SpeakInHouse->wData.val - 1]);
-    expForNextLevel = 1000ull * pParty->activeCharacter().uLevel * (pParty->activeCharacter().uLevel + 1) / 2;
-    //-------------------------------------------------------
-    all_text_height = 0;
-    if (HouseUI_CheckIfPlayerCanInteract()) {
-        if (dialog_menu_id == DIALOGUE_MAIN) {
-            if (HouseUI_CheckIfPlayerCanInteract()) {
-                index = 0;
-                pShopOptions[0] = s;  // set first item to fucntion param - this
-                                      // always gets overwritten below??
-                pShopOptions[1] = localization->GetString(LSTR_LEARN_SKILLS);
-                if (pDialogueWindow->pStartingPosActiveItem < pDialogueWindow->pStartingPosActiveItem + pDialogueWindow->pNumPresenceButton) {
-                    for (int i = pDialogueWindow->pStartingPosActiveItem; i < pDialogueWindow->pNumPresenceButton + pDialogueWindow->pStartingPosActiveItem; ++i) {
-                        if (pDialogueWindow->GetControl(i)->msg_param == DIALOGUE_TRAINING_HALL_TRAIN) {
-                            static std::string shop_option_str_container;
-                            if (pParty->activeCharacter().uLevel >= trainingHallMaxLevels[HOUSE_ID(window_SpeakInHouse->wData.val)]) {
-                                shop_option_str_container = fmt::format(
-                                    "{}\n \n{}",
-                                    localization->GetString(LSTR_TEACHER_LEVEL_TOO_LOW),
-                                    localization->GetString(LSTR_CANT_TRAIN_FURTHER));
-                                pShopOptions[index] = shop_option_str_container.c_str();
-                            } else {
-                                if (pParty->activeCharacter().experience < expForNextLevel)
-                                    shop_option_str_container = localization->FormatString(
-                                        LSTR_XP_UNTIL_NEXT_LEVEL,
-                                        (uint)(expForNextLevel - pParty->activeCharacter().experience),
-                                        pParty->activeCharacter().uLevel + 1);
-                                else
-                                    shop_option_str_container = localization->FormatString(
-                                        LSTR_FMT_TRAIN_LEVEL_D_FOR_D_GOLD,
-                                        pParty->activeCharacter().uLevel + 1,
-                                        pPrice);
-                                pShopOptions[index] = shop_option_str_container.c_str();
-                            }
-                        }
-                        all_text_height += pFontArrus->CalcTextHeight(pShopOptions[index], training_dialog_window.uFrameWidth, 0);
-                        ++index;
-                    }
-                }
-                v49 =
-                    (2 * (87 - (SIDE_TEXT_BOX_BODY_TEXT_HEIGHT - all_text_height) / 2) - all_text_height) /
-                    2 -
-                    (SIDE_TEXT_BOX_BODY_TEXT_HEIGHT - all_text_height) / 2 / 2 + SIDE_TEXT_BOX_BODY_TEXT_OFFSET;
-                if (pDialogueWindow->pStartingPosActiveItem <
-                    pDialogueWindow->pStartingPosActiveItem +
-                    pDialogueWindow->pNumPresenceButton) {
-                    index = 0;
-                    for (int i = pDialogueWindow->pStartingPosActiveItem;
-                        i < pDialogueWindow->pStartingPosActiveItem +
-                        pDialogueWindow->pNumPresenceButton;
-                        ++i) {
-                        pButton = pDialogueWindow->GetControl(i);
-                        pButton->uY = (SIDE_TEXT_BOX_BODY_TEXT_HEIGHT - all_text_height) / 2 + v49;
-                        pTextHeight = pFontArrus->CalcTextHeight(
-                            pShopOptions[index],
-                            training_dialog_window.uFrameWidth, 0);
-                        pButton->uHeight = pTextHeight;
-                        pButton->uW = pTextHeight + pButton->uY - 1 + 6;
-                        v49 = pButton->uW;
-                        int pTextColor = colorTable.Jonquil.c16();
-                        if (pDialogueWindow->pCurrentPosActiveItem != i)
-                            pTextColor = colorTable.White.c16();
-                        training_dialog_window.DrawTitleText(
-                            pFontArrus, 0, pButton->uY, pTextColor,
-                            pShopOptions[index], 3);
-                        ++index;
-                    }
-                }
-            }
-        }
-        //------------------------------------------------------------------
-        if (dialog_menu_id == DIALOGUE_TRAINING_HALL_TRAIN) {
-            std::string label;
-
-            if (!HouseUI_CheckIfPlayerCanInteract()) {
-                v33 = pFontArrus->CalcTextHeight(pNPCTopics[122].pText, training_dialog_window.uFrameWidth, 0);
-                training_dialog_window.DrawTitleText(pFontArrus, 0, (212 - v33) / 2 + 101, colorTable.Jonquil.c16(), pNPCTopics[122].pText, 3);
-                pDialogueWindow->pNumPresenceButton = 0;
-                return;
-            }
-            if (pParty->activeCharacter().uLevel < trainingHallMaxLevels[HOUSE_ID(window_SpeakInHouse->wData.val)]) {
-                if ((int64_t)pParty->activeCharacter().experience >= expForNextLevel) {
-                    if (pParty->GetGold() >= pPrice) {
-                        pParty->TakeGold(pPrice);
-                        PlayHouseSound(window_SpeakInHouse->wData.val, HouseSound_NotEnoughMoney);
-                        pParty->activeCharacter().uLevel++;
-                        pParty->activeCharacter().uSkillPoints += pParty->activeCharacter().uLevel / 10 + 5;
-                        pParty->activeCharacter().health = pParty->activeCharacter().GetMaxHealth();
-                        pParty->activeCharacter().mana = pParty->activeCharacter().GetMaxMana();
-                        int maxLevelStepsBefore = *std::max_element(charactersTrainedLevels.begin(), charactersTrainedLevels.end());
-                        charactersTrainedLevels[pParty->activeCharacterIndex() - 1]++;
-                        int maxLevelStepsAfter = *std::max_element(charactersTrainedLevels.begin(), charactersTrainedLevels.end());
-                        if (maxLevelStepsAfter > maxLevelStepsBefore) {
-                            GameTime trainingTime = GameTime::FromHours(_494820_training_time(pParty->uCurrentHour) + 4).SubtractMinutes(pParty->uCurrentMinute);
-                            if (window_SpeakInHouse->wData.val == HOUSE_TRAINING_HALL_PIT || window_SpeakInHouse->wData.val == HOUSE_TRAINING_HALL_NIGHON) {
-                                trainingTime.AddHours(12);
-                            }
-                            restAndHeal(trainingTime.AddDays(7));
-                            if (uCurrentlyLoadedLevelType == LEVEL_Outdoor)
-                                pOutdoor->SetFog();
-                        }
-                        pParty->activeCharacter().playReaction(SPEECH_LevelUp);
-
-                        GameUI_SetStatusBar(
-                            LSTR_FMT_S_NOW_LEVEL_D,
-                            pParty->activeCharacter().name.c_str(),
-                            pParty->activeCharacter().uLevel,
-                            pParty->activeCharacter().uLevel / 10 + 5
-                        );
-
-                        pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-                        return;
-                    }
-
-                    GameUI_SetStatusBar(LSTR_NOT_ENOUGH_GOLD);
-                    PlayHouseSound(window_SpeakInHouse->wData.val, (HouseSoundID)4);
-                    pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-                    return;
-                }
-                label = localization->FormatString(
-                    LSTR_XP_UNTIL_NEXT_LEVEL,
-                    (unsigned int)(expForNextLevel - pParty->activeCharacter().experience),
-                    pParty->activeCharacter().uLevel + 1);
-                v36 = (212 - pFontArrus->CalcTextHeight(
-                        label, training_dialog_window.uFrameWidth, 0)) / 2 + 88;
-            } else {
-                label = fmt::format(
-                    "{}\n \n{}",
-                    localization->GetString(LSTR_TEACHER_LEVEL_TOO_LOW),
-                    localization->GetString(LSTR_CANT_TRAIN_FURTHER));
-                v36 = (212 - pFontArrus->CalcTextHeight(
-                        label, training_dialog_window.uFrameWidth, 0)) / 2 + 101;
-            }
-            training_dialog_window.DrawTitleText(
-                pFontArrus, 0, v36, colorTable.Jonquil.c16(), label, 3);
-
-            PlayHouseSound(window_SpeakInHouse->wData.val, (HouseSoundID)3);
-            pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
-            return;
-        }
-    }
-
-    //-------------------------------------------------------------
-    if (dialog_menu_id == DIALOGUE_LEARN_SKILLS) {
-        if (HouseUI_CheckIfPlayerCanInteract()) {
-            pPrice = PriceCalculator::skillLearningCostForPlayer(&pParty->activeCharacter(),
-                                                                 buildingTable[window_SpeakInHouse->wData.val - 1]);
-            index = 0;
-            for (int i = pDialogueWindow->pStartingPosActiveItem;
-                (signed int)i < pDialogueWindow->pNumPresenceButton +
-                pDialogueWindow->pStartingPosActiveItem;
-                ++i) {
-                auto skill = GetLearningDialogueSkill(
-                    (DIALOGUE_TYPE)pDialogueWindow->GetControl(i)->msg_param
-                );
-                if (skillMaxMasteryPerClass[pParty->activeCharacter().classType][skill] != PLAYER_SKILL_MASTERY_NONE
-                    && !pParty->activeCharacter().pActiveSkills[skill]) {
-                    all_text_height += pFontArrus->CalcTextHeight(
-                        localization->GetSkillName(skill),
-                        training_dialog_window.uFrameWidth, 0);
-                    ++index;
-                }
-            }
-
-            SkillTrainingDialogue(&training_dialog_window, index, all_text_height, pPrice);
-        }
-    }
-    return;
-}
-
-/**
-  * @offset 0x004B6478.
-  *
-  * @brief Mercenary guild in MM6.
-  */
-void MercenaryGuildDialog() {
-    signed int v3;                // esi@1
-    short *v6;                       // edi@6
-    int all_text_height;          // eax@20
-    int pTextHeight;              // eax@29
-    int v27;                      // [sp-4h] [bp-80h]@8
-    int v32;                      // [sp+6Ch] [bp-10h]@1
-    int index;                    // [sp+74h] [bp-8h]@17
-
-    GUIWindow dialog_window = *window_SpeakInHouse;
-    dialog_window.uFrameX = SIDE_TEXT_BOX_POS_X;
-    dialog_window.uFrameWidth = SIDE_TEXT_BOX_WIDTH;
-    dialog_window.uFrameZ = SIDE_TEXT_BOX_POS_Z;
-
-    /*
-     * archiving this code just in case
-     * I believe it is 250 gold cost for mercenary guild from mm6 and 100 for all other skill-learning house types in mm6
-     * but they aren't used in mm7, so I'm gonna assume 250 gold cost in price calculator
-     *
-     *  v32 = (uint8_t)(((buildingTable[window_SpeakInHouse->wData.val - 1].uType != BuildingType_MercenaryGuild) - 1) & 0x96) + 100;
-     *  v3 = (int64_t)((double)v32 * buildingTable[window_SpeakInHouse->wData.val - 1].fPriceMultiplier);
-     *  pPrice = v3 * (100 - PriceCalculator::playerMerchant(&pParty->activeCharacter())) / 100;
-     *  if (pPrice < v3 / 3) pPrice = v3 / 3;
-     */
-    int pPrice = PriceCalculator::skillLearningCostForPlayer(&pParty->activeCharacter(), buildingTable[window_SpeakInHouse->wData.val - 1]);
-
-    if (dialog_menu_id == DIALOGUE_MAIN) {
-        if (!pParty->activeCharacter()._achievedAwardsBits[word_4F0754[2 * window_SpeakInHouse->wData.val]]) {
-            // 171 looks like Mercenary Stronghold message from NPCNews.txt in MM6
-            pTextHeight = pFontArrus->CalcTextHeight(pNPCTopics[171].pText, dialog_window.uFrameWidth, 0);
-            dialog_window.DrawTitleText(pFontArrus, 0, (212 - pTextHeight) / 2 + 101, colorTable.PaleCanary.c16(), pNPCTopics[171].pText, 3);
-            pDialogueWindow->pNumPresenceButton = 0;
-            return;
-        }
-        if (!HouseUI_CheckIfPlayerCanInteract()) return;
-        all_text_height = 0;
-        index = 0;
-        for (int i = pDialogueWindow->pStartingPosActiveItem; i < pDialogueWindow->pNumPresenceButton + pDialogueWindow->pStartingPosActiveItem; ++i) {
-            auto skill = GetLearningDialogueSkill((DIALOGUE_TYPE)pDialogueWindow->GetControl(i)->msg_param);
-            // Was class type / 3
-            if (skillMaxMasteryPerClass[pParty->activeCharacter().classType][skill] != PLAYER_SKILL_MASTERY_NONE
-                && !pParty->activeCharacter().pActiveSkills[skill]) {
-                all_text_height += pFontArrus->CalcTextHeight(localization->GetSkillName(skill), dialog_window.uFrameWidth, 0);
-                ++index;
-            }
-        }
-
-
-        SkillTrainingDialogue(&dialog_window, index, all_text_height, pPrice);
-        return;
-    }
-
-    if (HouseUI_CheckIfPlayerCanInteract()) {
-        __debugbreak();  // what type of house that even is?
-        // pSkillAvailabilityPerClass[8 + v58->uClass][4 + v23]
-        // or
-        // skillMaxMasteryPerClass[v58->uClass][v23 - 36]
-        // or
-        // skillMaxMasteryPerClass[v58->uClass - 1][v23 +
-        // 1]
-        __debugbreak();  // whacky condition - fix
-        if (false
-            // if ( !*(&byte_4ED94C[37 * v1->uClass / 3] + dword_F8B19C)
-            || (v6 = (short *)(&pParty->activeCharacter().uIntelligence +
-                dialog_menu_id),
-                *(short *)v6)) {
-            pAudioPlayer->playUISound(SOUND_error);
-        } else {
-            if (pParty->GetGold() < pPrice) {
-                GameUI_SetStatusBar(LSTR_NOT_ENOUGH_GOLD);
-                v27 = 4;
-            } else {
-                pParty->TakeGold(pPrice);
-                *(short *)v6 = 1;
-                v27 = 2;
-            }
-            PlayHouseSound(window_SpeakInHouse->wData.val, (HouseSoundID)v27);
-        }
-    }
-    pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
 }
 
 void SimpleHouseDialog() {
@@ -2636,18 +1493,6 @@ void SimpleHouseDialog() {
             _591428_endcap);
         house_window.DrawText(pTextFont, {13, 354 - pTextHeight}, colorTable.Black.c16(), pTextFont->FitTextInAWindow(current_npc_text, w.uFrameWidth, 13), 0, 0, 0);
     }
-}
-
-void JailDialog() {
-    GUIWindow jail_dialogue_window = *window_SpeakInHouse;
-    jail_dialogue_window.uFrameX = SIDE_TEXT_BOX_POS_X;
-    jail_dialogue_window.uFrameWidth = SIDE_TEXT_BOX_WIDTH;
-    jail_dialogue_window.uFrameZ = SIDE_TEXT_BOX_POS_Z;
-    jail_dialogue_window.DrawTitleText(
-        pFontArrus, 0,
-        (310 - pFontArrus->CalcTextHeight(localization->GetString(LSTR_ONE_YEAR_SENTENCE),
-            jail_dialogue_window.uFrameWidth,
-            0)) / 2 + 18, colorTable.PaleCanary.c16(), localization->GetString(LSTR_ONE_YEAR_SENTENCE), 3);
 }
 
 //----- (004BE571) --------------------------------------------------------
